+++ resources/conf/api.json
@@ -1,0 +1,4 @@
+{
+    "urlBaseApi": "http://localhost:5000",
+    "urlBaseBancoDados": "http://localhost:3333"
+}
