+++ resources/conf/web.json
@@ -1,0 +1,3 @@
+{
+    "urlBase": "http://localhost:3000"
+}
